Update zsh config.

+++ zsh/lib/zle.zsh
@@ -1,6 +1,6 @@
 slash-backward-kill-word() {
-    local WORDCHARS="${WORDCHARS:s@/@}"
-    zle backward-kill-word
+  local WORDCHARS="${WORDCHARS:s@/@}"
+  zle backward-kill-word
 }
 zle -N slash-backward-kill-word
 # <alt+backspace>
@@ -24,20 +24,27 @@ zle -N _my_clear
 bindkey '^l' _my_clear
 
 
-# Ctrl-O opens zsh at the current location, and on exit, cd into ranger's last location.
-ranger-cd() {
+if (( $+commands[ranger] )); then
+  # Ctrl-O opens zsh at the current location, and on exit, cd into ranger's last location.
+  ranger-cd() {
   tempfile=$(mktemp)
   ranger --choosedir="$tempfile" "${@:-$(pwd)}" < $TTY
   test -f "$tempfile" &&
-  if [ "$(cat -- "$tempfile")" != "$(echo -n `pwd`)" ]; then
-    cd -- "$(cat "$tempfile")"
-  fi
-  rm -f -- "$tempfile"
-  # hacky way of transferring over previous command and updating the screen
-  VISUAL=true zle edit-command-line
-}
-zle -N ranger-cd
-bindkey '^o' ranger-cd
+    if [ "$(cat -- "$tempfile")" != "$(echo -n `pwd`)" ]; then
+      cd -- "$(cat "$tempfile")"
+    fi
+    rm -f -- "$tempfile"
+    # hacky way of transferring over previous command and updating the screen
+    VISUAL=true zle edit-command-line
+  }
+  zle -N ranger-cd
+  bindkey '^o' ranger-cd
+fi
+
+bindkey -s '\em' '^Umycli -u root^M'
+if (( $+commands[mycli] )); then
+  bindkey -s '\em' '^Umycli -u root^M'
+fi
 
 goto-gitroot() {
   cd $(git rev-parse --show-toplevel 2> /dev/null || echo -n ".")
@@ -49,7 +56,6 @@ bindkey '^g' goto-gitroot
 autosuggest-accept-line() {
   zle autosuggest-accept
   zle accept-line
-
 }
 zle -N autosuggest-accept-line
 bindkey '^j' autosuggest-accept-line
